+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- User option to specify default editor layout ([#3](https://github.com/benji300/joplin-persistent-layout/issues/3))
 - Support to persist rich text editor layout ([#4](https://github.com/benji300/joplin-persistent-layout/issues/4))
 
 ## [1.0.0] - 2021-01-29